Get block decoration working sorta
