+++ packages/context-editor/src/plugins/index.ts
@@ -2,7 +2,7 @@ import { mathPlugin } from "@benrbray/prosemirror-math";
 import { exampleSetup } from "prosemirror-example-setup";
 import { Schema } from "prosemirror-model";
 
-import type { ContextEditorProps, PanelProps } from "../ContextEditor";
+import type { ContextEditorProps } from "../ContextEditor";
 import attributePanel from "./attributePanel";
 import code from "./code";
 import contextSuggest from "./contextSuggest";
@@ -16,8 +16,6 @@ import structureDecorations from "./structureDecorations";
 export const basePlugins = (
 	schema: Schema,
 	props: ContextEditorProps,
-	panelPosition: PanelProps,
-	setPanelPosition: React.Dispatch<React.SetStateAction<PanelProps>>,
 	suggestData: any,
 	setSuggestData: any
 ) => {
@@ -27,10 +25,9 @@ export const basePlugins = (
 		// Example setup includes inputRules for headers, blockquotes, codeblock, lists
 		// https://github.com/ProseMirror/prosemirror-example-setup/blob/master/src/inputrules.ts
 		...exampleSetup({ schema, menuBar: false }),
-		// reactProps(props),
-		// structureDecorations(),
-		// attributePanel(panelPosition, setPanelPosition),
-		// onChange(),
+		reactProps(props),
+		structureDecorations(),
+		onChange(),
 		pasteRules(schema),
 		inputRules(schema),
 		mathPlugin,
